21.11.25 ========  OpenGL on D3D11  ---------------   + New system translates OpenGL display output to D3D11 to make GL games       compatible with windowed mode G-Sync, Reflex and HDR (SK or AutoHDR)
    Feature is enabled by default, but can be disabled by setting:

      [API.Hook]
      d3d11=false

    Turning off D3D11 support in Special K will use legacy OpenGL

+++ src/render/gl/opengl_screenshot.cpp
@@ -108,8 +108,8 @@ SK_GL_Screenshot::SK_GL_Screenshot (const HGLRC hDevice) : hglrc (hDevice)
     static SK_RenderBackend& rb =
       SK_GetCurrentRenderBackend ();
 
-    framebuffer.Width        = io.DisplaySize.x;
-    framebuffer.Height       = io.DisplaySize.y;
+    framebuffer.Width        = static_cast <UINT> (io.DisplaySize.x);
+    framebuffer.Height       = static_cast <UINT> (io.DisplaySize.y);
     framebuffer.NativeFormat = DXGI_FORMAT_R8G8B8A8_UNORM;
 
     DirectX::ComputePitch (
